DTO methods
